[refactor]: 보관함 폴더/아이템 서비스 네이밍 분리를 위한 수정
- 보관함 폴더 관련: folder
- 보관함 아이템 관련: item
