+++ frontend/src/api/endpoints.ts
@@ -70,5 +70,8 @@ export const ENDPOINTS = {
       CREATE: '/api/v2/inventory/folder',
       DELETE: (id: number) => `/api/v2/inventory/folder/${id}`,
     },
+    ITEM: {
+      LIST: (id: number) => `/api/v2/inventory/${id}`,
+    },
   },
 };
