+++ frontend/src/hooks/queries/inventory/folder/useDeleteInventoryFolder.ts
@@ -1,11 +1,11 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { inventoryService } from '@/services/inventoryService';
+import { inventoryFolderService } from '@/services/inventoryFolderService';
 
 const useDeleteInventoryFolder = () => {
   const queryClient = useQueryClient();
 
   const { mutate, isPending, isError, error, data, reset } = useMutation({
-    mutationFn: (id: number) => inventoryService.delete(id),
+    mutationFn: (id: number) => inventoryFolderService.delete(id),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['inventoryFolderList'] });
     },
